Fixed offset issues

+++ src/jpower-apps/JPower.Shared/JPower/IJPowerDevice.cs
@@ -66,7 +66,7 @@ namespace JPower.Shared.JPower
         [MarshalAs(UnmanagedType.ByValArray, SizeConst = 16)]
         [FieldOffset(0)] public byte[] guid;
         [FieldOffset(16)] public float crankLength;
-        [FieldOffset(20)] public ushort offset;
+        [FieldOffset(20)] public uint offset;
         [FieldOffset(24)] public float slope;
         [FieldOffset(28)] public float intercept;
     }
